+++ packages/signers/meta-signers/src/wrapped-tx-creator.ts
@@ -1,4 +1,4 @@
-import { TxCreator } from "@polkadot-api/polkadot-signer"
+import { TxCreator } from "@polkadot-api/tx-creator"
 
 export type WrapTxCreator<T extends TxCreator> = {
   publicKey: Uint8Array