+++ packages/client/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+### Changed
+
+- BREAKING: Renamed `polkadot-api/signer` to `polkadot-api/tx-creator`
+
 ## 3.0.0-rc.2 - 2026-07-13
 
 ### Added